requirements.txt
flask==3.1.0
flask-cors==5.0.1
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1
numpy==2.2.2
pandas==2.2.3
groq==0.15.0
gunicorn==23.0.0
